feat(Facade): service 계층 LoginProfile 인가 기능 추가

+++ backend/src/main/java/com/cruru/process/service/facade/ProcessFacade.java
@@ -1,13 +1,17 @@
 package com.cruru.process.service.facade;
 
+import com.cruru.advice.ForbiddenException;
 import com.cruru.applicant.controller.dto.ApplicantCardResponse;
 import com.cruru.applicant.domain.Applicant;
 import com.cruru.applicant.service.ApplicantService;
 import com.cruru.applyform.domain.ApplyForm;
 import com.cruru.applyform.service.ApplyFormService;
+import com.cruru.auth.controller.dto.LoginProfile;
 import com.cruru.dashboard.domain.Dashboard;
 import com.cruru.dashboard.service.DashboardService;
 import com.cruru.evaluation.service.EvaluationService;
+import com.cruru.member.domain.Member;
+import com.cruru.member.service.MemberService;
 import com.cruru.process.controller.dto.ProcessCreateRequest;
 import com.cruru.process.controller.dto.ProcessResponse;
 import com.cruru.process.controller.dto.ProcessResponses;
@@ -24,6 +28,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class ProcessFacade {
 
+    private final MemberService memberService;
     private final ProcessService processService;
     private final DashboardService dashboardService;
     private final ApplicantService applicantService;
@@ -31,13 +36,26 @@ public class ProcessFacade {
     private final ApplyFormService applyFormService;
 
     @Transactional
-    public void create(ProcessCreateRequest request, long dashboardId) {
+    public void create(LoginProfile loginProfile, ProcessCreateRequest request, long dashboardId) {
+        Member member = memberService.findByEmail(loginProfile.email());
         Dashboard dashboard = dashboardService.findById(dashboardId);
+
+        validateOwner(dashboard, member);
         processService.create(request, dashboard);
     }
 
-    public ProcessResponses readAllByDashboardId(long dashboardId) {
+    private void validateOwner(Dashboard dashboard, Member member) {
+        if (!dashboard.isOwner(member)) {
+            throw new ForbiddenException();
+        }
+    }
+
+    public ProcessResponses readAllByDashboardId(LoginProfile loginProfile, long dashboardId) {
+        Member member = memberService.findByEmail(loginProfile.email());
         Dashboard dashboard = dashboardService.findById(dashboardId);
+
+        validateOwner(dashboard, member);
+
         ApplyForm applyForm = applyFormService.findByDashboard(dashboard);
         List<Process> processes = processService.findAllByDashboard(dashboard);
         List<ProcessResponse> processResponses = toProcessResponses(processes);
@@ -77,14 +95,25 @@ public class ProcessFacade {
     }
 
     @Transactional
-    public ProcessResponse update(ProcessUpdateRequest request, long processId) {
+    public ProcessResponse update(LoginProfile loginProfile, ProcessUpdateRequest request, long processId) {
+        Member member = memberService.findByEmail(loginProfile.email());
         Process process = processService.findById(processId);
+        validateOwner(process, member);
         processService.update(request, process.getId());
         return toProcessResponse(process);
     }
 
+    private void validateOwner(Process process, Member member) {
+        if (!process.isOwner(member)) {
+            throw new ForbiddenException();
+        }
+    }
+
     @Transactional
-    public void delete(long processId) {
+    public void delete(LoginProfile loginProfile, long processId) {
+        Member member = memberService.findByEmail(loginProfile.email());
+        Process process = processService.findById(processId);
+        validateOwner(process, member);
         processService.delete(processId);
     }
 }
